Create `BroadcastBundleInfo`
Encapsulates settings related to broadcast extension.
