+++ Sources/LiveKit/Broadcast/BroadcastScreenCapturer.swift
@@ -36,9 +36,7 @@ class BroadcastScreenCapturer: BufferCapturer {
 
         guard didStart else { return false }
 
-        guard let groupIdentifier = Self.groupIdentifier,
-              let socketPath = Self.socketPath(for: groupIdentifier)
-        else {
+        guard let socketPath = BroadcastBundleInfo.socketPath else {
             logger.error("Bundle settings improperly configured for screen capture")
             return false
         }
@@ -83,29 +81,6 @@ class BroadcastScreenCapturer: BufferCapturer {
         frameReader?.stopCapture()
         frameReader = nil
         return true
-    }
-
-    /// Identifier of the app group shared by the primary app and broadcast extension.
-    @BundleInfo("RTCAppGroupIdentifier")
-    static var groupIdentifier: String?
-
-    /// Bundle identifier of the broadcast extension.
-    @BundleInfo("RTCScreenSharingExtension")
-    static var screenSharingExtension: String?
-
-    /// Path to the socket file used for interprocess communication.
-    static var socketPath: String? {
-        guard let groupIdentifier = Self.groupIdentifier else { return nil }
-        return Self.socketPath(for: groupIdentifier)
-    }
-
-    private static let kRTCScreensharingSocketFD = "rtc_SSFD"
-
-    private static func socketPath(for groupIdentifier: String) -> String? {
-        guard let sharedContainer = FileManager.default
-            .containerURL(forSecurityApplicationGroupIdentifier: groupIdentifier)
-        else { return nil }
-        return sharedContainer.appendingPathComponent(Self.kRTCScreensharingSocketFD).path
     }
 }
